Correct firebase configuration and css
