+++ apps/merchant-web/src/app/globals.css
@@ -2,7 +2,6 @@
 @tailwind components;
 @tailwind utilities;
 
-
 :root {
   --background: #ffffff;
   --foreground: #171717;
@@ -22,25 +21,15 @@ body {
 }
 
 body {
-  padding: 25px;
-  height: max-content;
   color: var(--foreground);
   background: var(--background);
-  font-family: Inter, Arial, Helvetica, sans-serif;
+  font-family: Arial, Helvetica, sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-
-  color: #000;
-  background-color: #F4F9FF;
-}
-
-[data-theme="dark"] body {
-  background-color: #282828;
-  color: #fff;
 }
 
 * {
-  /*box-sizing: border-box;*/
+  box-sizing: border-box;
   padding: 0;
   margin: 0;
 }
@@ -50,13 +39,8 @@ a {
   text-decoration: none;
 }
 
-/* @media (prefers-color-scheme: dark) {
+@media (prefers-color-scheme: dark) {
   html {
     color-scheme: dark;
   }
-}*/
-
-form {
-  margin-top: 10px;
-  margin-bottom: 10px;
 }
