+++ apps/merchant-web/src/styles/admin-panel/add-products.module.css
@@ -1,3 +1,7 @@
+.addProductForm {
+    padding: 25px;
+}
+
 .heading1 {
     font-size: 28px;
     color: #7790ED;
